[Test-case3 to 8] Adding discount affecting price

+++ src/main/kotlin/Delivery.kt
@@ -1,6 +1,6 @@
 class Delivery(
     val id: Int,
     val name: String,
-    val discountPercent: Int,
+    val discountPercent: Double,
     val price: Double,
 )
